Refactor: 폴더 경로 정리
