+++ src/routes/awsUpload.ts
@@ -1,5 +1,5 @@
 import express, { Request, Response } from "express";
-import { s3, upload } from "../s3Uploader";
+import { s3, upload } from "../../config/s3Uploader";
 import { PutObjectCommand, PutObjectCommandInput } from "@aws-sdk/client-s3";
 import { v4 as uuidv4 } from "uuid";
